move metastore to own file
